Fix safety issues, consolidate constants, remove dead code
- Fix IndexError on empty displayName in profile parser
- Add _get_csrf_token() with clear error on missing CSRF cookie
- Consolidate constants: remove duplicates between const.py files
- Fix MANUFACTURER typo (was MANUFACTRURER)
- Remove stale API_VERSION="19" from top-level const.py
- Add comment explaining folder_id mapping in notification parser
- Remove unused services.yaml (service was never registered)
