+++ custom_components/aula/aula_proxy/const.py
@@ -1,18 +1,3 @@
-STARTUP = r"""
-                _
-     /\        | |
-    /  \  _   _| | __ _
-   / /\ \| | | | |/ _` |
-  / ____ \ |_| | | (_| |
- /_/    \_\__,_|_|\__,_|
-Aula integration, version: %s
-This is a custom integration
-If you have any issues with this you need to open an issue here:
-https://github.com/scaarup/aula/issues
--------------------------------------------------------------------
-"""
-
-DOMAIN = "aula"
 API = "https://www.aula.dk/api/v"
 API_VERSION = "23"
 MIN_UDDANNELSE_API = "https://api.minuddannelse.net/aula"
