+++ custom_components/aula/const.py
@@ -12,13 +12,5 @@ https://github.com/scaarup/aula/issues
 -------------------------------------------------------------------
 """
 
-MANUFACTRURER = "Aula"
+MANUFACTURER = "Aula"
 DOMAIN = "aula"
-API = "https://www.aula.dk/api/v"
-API_VERSION = "19"
-MIN_UDDANNELSE_API = "https://api.minuddannelse.net/aula"
-MEEBOOK_API = "https://app.meebook.com/aulaapi"
-SYSTEMATIC_API = "https://systematic-momo.dk/api/aula"
-EASYIQ_API = "https://api.easyiqcloud.dk/api/aula"
-# CONF_SCHOOLSCHEDULE = "schoolschedule"
-# CONF_UGEPLAN = "ugeplan"
